build: 打包成lib,h5,taro3个包
