+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.0.40",
   "description": "腾讯连连开发面板组件",
   "main": "lib/index.js",
+  "browser": "h5/index.js",
   "typings": "lib/index.d.ts",
   "module": "es/index.js",
   "scripts": {
@@ -24,7 +25,8 @@
     "lib/**/*.js",
     "lib/**/*.css",
     "lib/**/*.d.ts",
-    "es/*"
+    "h5/*",
+    "taro/*"
   ],
   "homepage": "https://github.com/tencentyun/iot-panel-component",
   "bugs": {
@@ -57,6 +59,7 @@
     "rollup-plugin-import-css": "^3.0.0",
     "rollup-plugin-multi-input": "^1.3.1",
     "rollup-plugin-postcss": "^4.0.1",
+    "@rollup/plugin-replace": "^2.2.0",
     "rollup-plugin-static-import": "^0.1.1",
     "typescript": "^4.4.3"
   },
