+++ src/components/ScrollView/ScrollViewTaro.tsx
@@ -1,0 +1,1 @@
+export { ScrollView as default } from '@tarojs/components';
