+++ src/components/Slider/SliderTaro.tsx
@@ -1,0 +1,1 @@
+export { Slider as default } from '@tarojs/components';
