+++ src/components/Slider/index.ts
@@ -1,0 +1,1 @@
+export const Slider = process.env.TARO_ENV === 'weapp' ? require('./SliderTaro') : require('./SliderH5');
