+++ src/components/Switch/SwitchTaro.tsx
@@ -1,0 +1,1 @@
+export { Switch as default } from '@tarojs/components';
